% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootsSE.R
\name{bootsSE}
\alias{bootsSE}
\title{Standard error estimates by bootstrap sampling}
\usage{
bootsSE(
  object,
  nboots = 100,
  print.bootspara = TRUE,
  maxiter = 1000,
  ncores = 2,
  quadpoint = 10
)
}
\arguments{
\item{object}{an object inheriting from class JMM.}

\item{nboots}{number of bootstrap samples for standard error estimation. 
Default is 100.}

\item{print.bootspara}{logical; if TRUE a matrix of parameter estimates from 
all bootstrap samples will be returned.}

\item{maxiter}{number of iterations for EM algorithm. Default is 1000.}

\item{ncores}{number of cores to proceed parallel computation.}

\item{quadpoint}{number of Gauss-Hermite quadrature points used for numerical integration.}
}
\description{
Standard error estimates by bootstrap sampling
}
