% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fittedval.R
\name{fittedval}
\alias{fittedval}
\title{Get fitted values for multiple markers over time}
\usage{
fittedval(object, long.covar = NULL)
}
\arguments{
\item{object}{a object inheriting from class JMM.}

\item{long.covar}{a matrix of longitudinal covariates for biomarkers. 
Each row of observations stand for a particuler biomarker. 
If NULL the mean values of covariates in the longitudianl data will be adopted. Default is NULL.}
}
\description{
Get fitted values for multiple markers over time
}
