% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfit.R
\name{Simfit}
\alias{Simfit}
\title{Simulation results}
\usage{
Simfit(
  sim = 100,
  n = 215,
  tL = 0,
  tU = 9,
  nbreak = 8,
  p01 = 1,
  p02 = 1,
  j_max = 2,
  quadpoint = 10,
  maxiter = 4000,
  distr = "Gaussian",
  m_age = 0,
  std_age = 2,
  k_max = 2,
  q_eta = 1,
  t_max = 19,
  tsigmau_inv = NULL,
  tnbreak = 8,
  tbtheta = NULL,
  tbeta0 = NULL,
  tbeta1 = NULL,
  tsigma = NULL,
  tsigmad = NULL,
  ttheta = NULL,
  teta = NULL,
  tgamma = NULL,
  sigmau_invinit = NULL,
  bthetainit = NULL,
  beta0init = NULL,
  beta1init = NULL,
  sigmainit = NULL,
  thetainit = NULL,
  sigmadinit = NULL,
  gammainit = NULL,
  survVar = FALSE,
  lambda0 = NULL,
  out.bs = FALSE,
  conversigmad = FALSE,
  ncores = 2
)
}
\arguments{
\item{sim}{number of simulated datasets to be generated.}

\item{n}{sample size for each dataset.}

\item{tL}{lower limit of time point for spline basis.}

\item{tU}{upper limit of time point for spline basis.}

\item{nbreak}{number of knots for spline basis for a model fit.}

\item{p01}{number of covariates in the first biomarker.}

\item{p02}{number of covariates in the second biomarker.}

\item{j_max}{Total number of biomarkers in the longitudinal data.}

\item{quadpoint}{Number of quadrature points.}

\item{maxiter}{The maximum number of iterations for EM process.}

\item{distr}{distributional assumption of random effects.}

\item{m_age}{true mean value of age variable in the longitudinal sub-model.}

\item{std_age}{true standard deviation of age variable in the longitudinal sub-model.}

\item{k_max}{dimension of random effects.}

\item{q_eta}{number of survival covariates.}

\item{t_max}{maximum number of repeated measurements allowed for each subject.}

\item{tnbreak}{number of knots for spline basis for data generation.}

\item{tbtheta}{true btheta matrix.}

\item{tbeta0}{true value of beta0 matrix.}

\item{tbeta1}{true value of beta1 vector.}

\item{tsigma}{true value of sigma square vector.}

\item{tsigmad}{true value of variance of random effects.}

\item{ttheta}{true value of theta vector.}

\item{teta}{true value of survival fixed effects.}

\item{tgamma}{true value of latent association parameter.}

\item{bthetainit}{a data frame of the normalizing constant for spline basis.}

\item{beta0init}{Initial values of beta0.}

\item{beta1init}{Initial values of beta1.}

\item{sigmainit}{Initial values of sigma^2.}

\item{thetainit}{Initial values of theta.}

\item{sigmadinit}{Initial values of sigmad.}

\item{gammainit}{Initial values of gamma.}

\item{survVar}{logical; TRUE if the survival sub-model include covariates. Default is FALSE.}

\item{lambda0}{true baseline hazard.}

\item{out.bs}{logical; TRUE if output baseline hazard. Default is FALSE.}

\item{conversigmad}{logical; TRUE if sigmad is considered into convergence criteria. Default is FALSE.}

\item{ncores}{number of cores to proceed parallel computation.}

\item{tsigma_inv}{true sigma_inv matrix.}

\item{sigma_invinit}{Initial values of sigma_inv matrix.}
}
\description{
Simulation results
}
\seealso{
\code{\link{SimData}, \link{jmspline}}
}
