% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmspline.R
\name{jmspline}
\alias{jmspline}
\title{Joint Modelling for Continuous outcomes}
\usage{
jmspline(
  ydata,
  cdata,
  mdata,
  sigmau_inv,
  tbtheta,
  tL,
  tU,
  nbreak,
  p01,
  p02,
  j_max,
  k_max = 2,
  quadpoint = 10,
  maxiter = 2500,
  do.trace = FALSE,
  beta0init = NULL,
  beta1init = NULL,
  sigmainit = NULL,
  thetainit = NULL,
  sigmadinit = NULL,
  gammainit = NULL,
  survVar = TRUE
)
}
\arguments{
\item{ydata}{a multibiomarker longitudinal data frame in long format.}

\item{cdata}{a survival data frame with single failure.
Each subject has one data entry.}

\item{mdata}{a data vector of number of repeated measurements.}

\item{sigmau_inv}{sigmau_inv.}

\item{tbtheta}{a data frame of the normalizing constant for spline basis.}

\item{tL}{the lower limit of time interval for spline basis.}

\item{tU}{the upper limit of time interval for spline basis.}

\item{nbreak}{number of knots to be specified.}

\item{p01}{number of covariates in the first biomarker.}

\item{p02}{number of covariates in the second biomarker.}

\item{j_max}{Total number of biomarkers in the longitudinal data.}

\item{k_max}{Number of random effects to be specified. Default is 2.}

\item{quadpoint}{Number of quadrature points.}

\item{maxiter}{The maximum number of iterations for EM process.}

\item{do.trace}{A logic value to control the iterated results to be printed. Default is FALSE.}

\item{beta0init}{Initial values of beta0.}

\item{beta1init}{Initial values of beta1.}

\item{sigmainit}{Initial values of sigma^2.}

\item{thetainit}{Initial values of theta.}

\item{sigmadinit}{Initial values of sigmad.}

\item{gammainit}{Initial values of gamma.}

\item{survVar}{logical; TRUE if the survival sub-model include covariates. Default is TRUE.}
}
\description{
Joint modeling of longitudinal continuous and survival data
}
